apps/conversor-heif/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Image processing
Pillow==10.1.0
pillow-heif==0.14.0
PyMuPDF==1.23.8

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Utilities
python-dotenv==1.0.0
pathlib2==2.3.7
psutil==7.0.0
